services/ingestor_adsb/requirements.txt
httpx==0.27.*
asyncpg==0.29.*
redis==5.0.*
pyyaml==6.0.*
python-dotenv==1.0.*
